+++ viewer/src/main/java/config/TeamColorsPanel.java
@@ -16,6 +16,8 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTable;
 import javax.swing.border.Border;
+import javax.swing.event.TableModelEvent;
+import javax.swing.event.TableModelListener;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.table.TableCellEditor;
 import javax.swing.table.TableCellRenderer;
@@ -33,6 +35,8 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 	JTable teamColorTable;
 	DefaultTableModel tableModel;
 
+	private Runnable onChange;
+
 	public TeamColorsPanel(RVConfigure configProg)
 	{
 		super();
@@ -40,6 +44,7 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 		config = configProg.config.teamColors;
 		initGUI();
 		configProg.listeners.add(this);
+		onChange = configProg.updateSaveButtonState;
 	}
 
 	void initGUI()
@@ -95,6 +100,15 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 			i++;
 		}
 
+		tableModel.addTableModelListener(new TableModelListener() {
+			@Override
+			public void tableChanged(TableModelEvent e)
+			{
+				updateColorByTeamNameConfig();
+				onChange.run();
+			}
+		});
+
 		teamColorTable = new JTable(tableModel);
 		teamColorTable.setDefaultRenderer(Color.class, new ColorRenderer());
 		teamColorTable.setDefaultEditor(Color.class, new ColorEditor());
@@ -122,8 +136,7 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 			teamColorTable.setRowSelectionInterval(tableModel.getRowCount() - 1, tableModel.getRowCount() - 1);
 	}
 
-	@Override
-	public void configSaved(RVConfigure configProg)
+	private void updateColorByTeamNameConfig()
 	{
 		config.colorByTeamName.clear();
 		for (int i = 0; i < tableModel.getRowCount(); i++) {
@@ -131,6 +144,12 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 			Color color = (Color) tableModel.getValueAt(i, 1);
 			config.colorByTeamName.put(teamName, color);
 		}
+	}
+
+	@Override
+	public void configSaved(RVConfigure configProg)
+	{
+		updateColorByTeamNameConfig();
 	}
 
 	private static class TeamColorsTableModel extends DefaultTableModel
@@ -156,6 +175,7 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 			setOpaque(true);
 		}
 
+		@Override
 		public Component getTableCellRendererComponent(
 				JTable table, Object color, boolean isSelected, boolean hasFocus, int row, int column)
 		{
@@ -202,6 +222,7 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 			dialog.setIconImage(Globals.getIcon());
 		}
 
+		@Override
 		public void actionPerformed(ActionEvent e)
 		{
 			if (EDIT.equals(e.getActionCommand())) {
@@ -217,11 +238,13 @@ public class TeamColorsPanel extends JPanel implements SaveListener
 			}
 		}
 
+		@Override
 		public Object getCellEditorValue()
 		{
 			return currentColor;
 		}
 
+		@Override
 		public Component getTableCellEditorComponent(
 				JTable table, Object value, boolean isSelected, int row, int column)
 		{