+++ viewer/src/main/java/rv/Viewer.java
@@ -61,6 +61,7 @@ import rv.comm.rcssserver.scenegraph.SceneGraph;
 import rv.content.ContentManager;
 import rv.ui.UserInterface;
 import rv.ui.menus.MenuBar;
+import rv.util.MacEnhancements;
 import rv.util.commandline.Argument;
 import rv.util.commandline.BooleanArgument;
 import rv.util.commandline.IntegerArgument;
@@ -77,7 +78,7 @@ import rv.world.WorldModel;
 public class Viewer
 		extends GLProgram implements GLEventListener, ServerComm.ServerChangeListener, LogPlayer.StateChangeListener
 {
-	private static final String VERSION = "1.7.0";
+	private static final String VERSION = "1.8.0";
 
 	public enum Mode {
 		LOGFILE,
@@ -154,6 +155,7 @@ public class Viewer
 		return config;
 	}
 
+	@Override
 	public Viewport getScreen()
 	{
 		return screen;
@@ -227,8 +229,7 @@ public class Viewer
 		StringArgument drawingFilterArgument = new StringArgument("drawingFilter", ".*");
 
 		handleLogModeArgs(logFileArgument.parse(args), logModeArgument.parse(args));
-		config.networking.overrideServerHost(serverHostArgument.parse(args));
-		config.networking.overrideServerPort(serverPortArgument.parse(args));
+		config.networking.overrideServer(serverHostArgument.parse(args), serverPortArgument.parse(args));
 		drawingFilter = drawingFilterArgument.parse(args);
 		Argument.endParse(args);
 	}
@@ -260,7 +261,11 @@ public class Viewer
 
 	private void initComponents(GLCapabilities caps)
 	{
-		Globals.setLookFeel();
+		if (MacEnhancements.IS_MAC) {
+			MacEnhancements.useSystemMenuBar();
+		}
+
+		Globals.setLookFeel(config.general.lookAndFeel);
 		canvas = new GLCanvas(caps);
 		canvas.setFocusTraversalKeysEnabled(false);
 
@@ -307,8 +312,6 @@ public class Viewer
 
 	private void storeConfig()
 	{
-		Configuration config = Configuration.loadFromFile();
-
 		Configuration.Graphics graphics = config.graphics;
 		Point location = frame.getLocation();
 		Dimension size = frame.getSize();
@@ -376,11 +379,13 @@ public class Viewer
 		init = true;
 	}
 
+	@Override
 	public void addKeyListener(KeyListener l)
 	{
 		(new AWTKeyAdapter(l, canvas)).addTo(canvas);
 	}
 
+	@Override
 	public void addMouseListener(MouseListener l)
 	{
 		(new AWTMouseAdapter(l, canvas)).addTo(canvas);
@@ -431,6 +436,7 @@ public class Viewer
 		System.exit(1);
 	}
 
+	@Override
 	public void update(GL glGeneric)
 	{
 		if (!init)
@@ -504,8 +510,7 @@ public class Viewer
 		if (mode != Mode.LIVE)
 			return;
 
-		String host = server.isConnected() ? config.networking.getServerHost() + ":" + config.networking.getServerPort()
-										   : null;
+		String host = server.isConnected() ? server.getServerHost() + ":" + server.getServerPort() : null;
 		frame.setTitle(getTitle(host));
 	}